Fix an error when FA doesn't return correctly.
When uploading to FA and resizing, it was possible the page could
have had an error and the expected data would not be in a page.
Instead of throwing an error, it should now correctly show an
error message. This should also fix it not showing an error message
where it should have.